refac controller
