+++ ui/controller.go
@@ -137,10 +137,6 @@ func (c *controller) Listen(ctx context.Context, g *gocui.Gui, sub chan *events.
 	}
 }
 
-func quit(g *gocui.Gui, v *gocui.View) error {
-	return gocui.ErrQuit
-}
-
 func (c *controller) Help(g *gocui.Gui, v *gocui.View) error {
 	maxX, maxY := g.Size()
 	view := c.views.Get(g.CurrentView())
@@ -153,7 +149,7 @@ func (c *controller) Help(g *gocui.Gui, v *gocui.View) error {
 		return c.views.Help.Set(g, 0, -1, maxX, maxY)
 	}
 
-	err := g.DeleteView(views.HELP)
+	err := view.Delete(g)
 	if err != nil {
 		return err
 	}
@@ -318,70 +314,6 @@ func (c *controller) OnEnter(g *gocui.Gui, v *gocui.View) error {
 			return err
 		}
 	}
-	return nil
-}
-
-func (c *controller) setKeyBinding(g *gocui.Gui) error {
-	err := g.SetKeybinding("", gocui.KeyCtrlC, gocui.ModNone, quit)
-	if err != nil {
-		return err
-	}
-
-	err = g.SetKeybinding("", gocui.KeyF10, gocui.ModNone, quit)
-	if err != nil {
-		return err
-	}
-
-	err = g.SetKeybinding("", 'q', gocui.ModNone, quit)
-	if err != nil {
-		return err
-	}
-
-	err = g.SetKeybinding("", gocui.KeyArrowUp, gocui.ModNone, c.cursorUp)
-	if err != nil {
-		return err
-	}
-
-	err = g.SetKeybinding("", gocui.KeyArrowDown, gocui.ModNone, c.cursorDown)
-	if err != nil {
-		return err
-	}
-
-	err = g.SetKeybinding("", gocui.KeyArrowLeft, gocui.ModNone, c.cursorLeft)
-	if err != nil {
-		return err
-	}
-
-	err = g.SetKeybinding("", gocui.KeyArrowRight, gocui.ModNone, c.cursorRight)
-	if err != nil {
-		return err
-	}
-
-	err = g.SetKeybinding("", gocui.KeyEnter, gocui.ModNone, c.OnEnter)
-	if err != nil {
-		return err
-	}
-
-	err = g.SetKeybinding("", gocui.KeyF1, gocui.ModNone, c.Help)
-	if err != nil {
-		return err
-	}
-
-	err = g.SetKeybinding("", 'h', gocui.ModNone, c.Help)
-	if err != nil {
-		return err
-	}
-
-	err = g.SetKeybinding("", gocui.KeyF2, gocui.ModNone, c.Menu)
-	if err != nil {
-		return err
-	}
-
-	err = g.SetKeybinding("", 'm', gocui.ModNone, c.Menu)
-	if err != nil {
-		return err
-	}
-
 	return nil
 }
